backend/requirements.txt
# Core
fastapi==0.115.5
uvicorn[standard]==0.32.1
gunicorn==23.0.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Auth & Security
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
cryptography==44.0.0

# Validation
pydantic==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0

# Redis
redis[hiredis]==5.2.1

# HTTP client
httpx==0.28.1

# Payments
stripe==11.3.0

# Image processing
Pillow==11.0.0
python-multipart==0.0.18

# Email
aiosmtplib==3.0.2
jinja2==3.1.4

# SMS
africastalking==1.2.6

# Observability
python-json-logger==3.2.0

# Dev/Test
pytest==8.3.4
pytest-asyncio==0.25.0
